refactor(frontend): split the Mantine theme monolith into a theme module
