+++ packages/frontend/CLAUDE.md
@@ -12,7 +12,7 @@ the [Frontend Style Guide](../../.cursor/rules/frontend.mdc). Key points:
 -   **NEVER use** `styles`(v8) or `sx`(v6) props or `style`(v6/v8)
 -   **Colors**: Prefer default component colors (auto-theme switching). For custom colors, use `ldGray.X` and `ldDark.X`, not standard `gray.X`
 -   **Prop changes** - `spacing` → `gap`, `noWrap` → `wrap="nowrap"`, `sx` → `style` (v6)
--   **Shared layout vars** - Heights/widths/z-indexes (navbar, page content, sidebar, dashboard header/tabs) are global CSS vars sourced from `*/constants.ts` via `src/mantineCssVariablesResolver.ts`. Reference `var(--name)` in CSS; don't hardcode the literal or bridge it through an inline `style`. To add one: constant → resolver → `var()`.
+-   **Shared layout vars** - Heights/widths/z-indexes (navbar, page content, sidebar, dashboard header/tabs) are global CSS vars sourced from `*/constants.ts` via `src/theme/cssVariablesResolver.ts`. Reference `var(--name)` in CSS; don't hardcode the literal or bridge it through an inline `style`. To add one: constant → resolver → `var()`.
 -   **Component docs** - Props/APIs at `https://mantine.dev/core/[component-name]/` (e.g. select, segmented-control)
 
 ## 🧩 Reusable Components
