+++ packages/frontend/src/hooks/echarts/useEchartsSankeyConfig.ts
@@ -125,9 +125,7 @@ const useEchartsSankeyConfig = (isInDashboard?: boolean) => {
 
         return {
             textStyle: {
-                fontFamily: sanitizeEchartsFontFamily(
-                    theme?.other.chartFont as string | undefined,
-                ),
+                fontFamily: sanitizeEchartsFontFamily(theme?.other.chartFont),
             },
             tooltip: {
                 ...getTooltipStyle({ appendToBody: !isTouchDevice }),
