+++ packages/frontend/src/hooks/echarts/useEchartsTreemapConfig.ts
@@ -207,9 +207,7 @@ const useEchartsTreemapConfig = (isInDashboard: boolean) => {
 
         return {
             textStyle: {
-                fontFamily: sanitizeEchartsFontFamily(
-                    theme?.other?.chartFont as string | undefined,
-                ),
+                fontFamily: sanitizeEchartsFontFamily(theme?.other?.chartFont),
             },
             tooltip: {
                 ...getTooltipStyle({ appendToBody: !isTouchDevice }),
