+++ packages/frontend/src/providers/MantineProvider.tsx
@@ -2,7 +2,7 @@ import { type MantineThemeOverride } from '@mantine/core';
 import { useLocalStorage } from '@mantine/hooks';
 import { Notifications } from '@mantine/notifications';
 import { useEffect, useMemo, type FC } from 'react';
-import { type ColorScheme } from '../mantineTheme';
+import { type ColorScheme } from '../theme';
 import { ColorSchemeContext } from './ColorSchemeContext';
 import MantineBaseProvider from './MantineBaseProvider';
